安全性修改

+++ gulpfile.js
@@ -1,3 +1,4 @@
+const isWin = /^win/.test(process.platform);
 const fs = require('fs');
 const join = require('path').join;
 const execSync = require('child_process').execSync;
@@ -26,8 +27,8 @@ function findSync(startPath) {
     return result;
 }
 
-const checkDirExist = (folderpath,os) => {
-    const pathArr = folderpath.split(os);
+const checkDirExist = (folderpath) => {
+    const pathArr = folderpath.split(isWin?'\\':'/');
     let _path = '';
     for (let i = 0; i < pathArr.length; i++) {
         if (pathArr[i]) {
@@ -65,30 +66,32 @@ gulp.task('compress', async () => {
     gulp.src('src/main/**/*.min.js')
         .pipe(gulp.dest(buildBasePath));
 
-    // Closure Compiler-Win
-    // for (let i of findSync(__dirname + '/src/main')) {
-    //     i = i.replace(__dirname + '\\src\\main', '');
-    //     if (i.indexOf('config.json') < 0 && /^((?!js).*)js/.test(i) && !/^((?!min.js).*)min.js/.test(i)) {
-    //         let dUrl = __dirname + '\\dist' + i;
-    //         checkDirExist(dUrl.slice(0, dUrl.lastIndexOf('\\')));
-    //         if (fs.existsSync(dUrl)) fs.unlinkSync(dUrl);
-    //         let javaExe = "D:\\Program Files\\idea\\jbr\\bin\\java";
-    //         execSync(`"${javaExe}" -jar closure-compiler-v20200224.jar --js ${'src/main' + i} --js_output_file ${'dist' + i} --language_in=ECMASCRIPT_2017 --language_out=ECMASCRIPT_2017 --compilation_level=SIMPLE --jscomp_warning=* --env=CUSTOM --module_resolution=NODE`, {cwd: process.cwd()});
-    //     }
-    // }
 
-    // Closure Compiler-Linux
-    for (let i of findSync(__dirname + '/src/main')) {
-        i = i.replace(__dirname + '/src/main', '');
-        if (i.indexOf('config.json') < 0 && /^((?!js).*)js/.test(i) && !/^((?!min.js).*)min.js/.test(i)) {
-            let dUrl = __dirname + '/dist' + i;
-            checkDirExist(dUrl.slice(0, dUrl.lastIndexOf('/')),'/');
-            if (fs.existsSync(dUrl)) fs.unlinkSync(dUrl);
-            let javaExe = "/lib/idea/jbr/bin/java";
-            execSync(`"${javaExe}" -jar closure-compiler-v20200224.jar --js ${'src/main' + i} --js_output_file ${'dist' + i} --language_in=ECMASCRIPT_2017 --language_out=ECMASCRIPT_2017 --compilation_level=SIMPLE --jscomp_warning=* --env=CUSTOM --module_resolution=NODE`, {cwd: process.cwd()});
+    if (isWin){
+        // Closure Compiler-Win
+        for (let i of findSync(__dirname + '/src/main')) {
+            i = i.replace(__dirname + '\\src\\main', '');
+            if (i.indexOf('config.json') < 0 && /^((?!js).*)js/.test(i) && !/^((?!min.js).*)min.js/.test(i)) {
+                let dUrl = __dirname + '\\dist' + i;
+                checkDirExist(dUrl.slice(0, dUrl.lastIndexOf('\\')));
+                if (fs.existsSync(dUrl)) fs.unlinkSync(dUrl);
+                let javaExe = "D:\\Program Files\\idea\\jbr\\bin\\java";
+                execSync(`"${javaExe}" -jar closure-compiler-v20200224.jar --js ${'src/main' + i} --js_output_file ${'dist' + i} --language_in=ECMASCRIPT_2017 --language_out=ECMASCRIPT_2017 --compilation_level=SIMPLE --jscomp_warning=* --env=CUSTOM --module_resolution=NODE`, {cwd: process.cwd()});
+            }
+        }
+    }else {
+        // Closure Compiler-Linux
+        for (let i of findSync(__dirname + '/src/main')) {
+            i = i.replace(__dirname + '/src/main', '');
+            if (i.indexOf('config.json') < 0 && /^((?!js).*)js/.test(i) && !/^((?!min.js).*)min.js/.test(i)) {
+                let dUrl = __dirname + '/dist' + i;
+                checkDirExist(dUrl.slice(0, dUrl.lastIndexOf('/')));
+                if (fs.existsSync(dUrl)) fs.unlinkSync(dUrl);
+                let javaExe = "/lib/idea/jbr/bin/java";
+                execSync(`"${javaExe}" -jar closure-compiler-v20200224.jar --js ${'src/main' + i} --js_output_file ${'dist' + i} --language_in=ECMASCRIPT_2017 --language_out=ECMASCRIPT_2017 --compilation_level=SIMPLE --jscomp_warning=* --env=CUSTOM --module_resolution=NODE`, {cwd: process.cwd()});
+            }
         }
     }
-
 
     //html
     gulp.src('src/main/**/*.html')
